Fix Timeline/Project circular dependency
Inject Timeline in Project, and not vice-versa.
