+++ frontend/app/timelines/models/index.js
@@ -36,12 +36,12 @@ angular.module('openproject.timelines.models')
   .factory('PlanningElement', ['PathHelper', require('./planning-element')])
   .factory('ProjectAssociation', require('./project-association'))
   .factory('ProjectType', require('./project-type'))
-  .factory('Project', require('./project'))
+  .factory('Project', ['Timeline', require('./project')])
   .factory('Reporting', require('./reporting'))
   .factory('Status', require('./status'))
   .factory('Timeline', ['Constants', 'TreeNode', 'UI', 'Color',
     'HistoricalPlanningElement', 'PlanningElement', 'PlanningElementType',
-    'ProjectType', 'Project', 'ProjectAssociation', 'Reporting',
+    'ProjectType', 'ProjectAssociation', 'Reporting',
     'CustomField', 'CustomFieldHelper',
     require('./timeline')
   ])
